init autometa 2
