+++ autometa/config/default.cfg
@@ -1,0 +1,27 @@
+[common]
+home_dir = /Users/rees/Wisc/kwan/tools/autometa
+db_dir = ${common:home_dir}/databases
+ncbi_dir = ${common:db_dir}/ncbi
+markers_dir = ${common:db_dir}/markers
+
+[database_urls]
+taxdump = ftp://ftp.ncbi.nlm.nih.gov/pub/taxonomy/taxdump.tar.gz
+accession2taxid = ftp://ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/prot.accession2taxid.gz
+nr = ftp://ftp.ncbi.nlm.nih.gov/blast/db/FASTA/nr.gz
+
+[checksums]
+acc2taxid = ftp://ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/prot.accession2taxid.gz.md5
+taxdump = ftp://ftp.ncbi.nlm.nih.gov/pub/taxonomy/taxdump.tar.gz.md5
+
+[ncbi]
+names = ${common:ncbi_dir}/names.dmp
+nodes = ${common:ncbi_dir}/nodes.dmp
+accession2taxid = ${common:ncbi_dir}/prot.accession2taxid.gz
+blastdb = ${common:ncbi_dir}/nr.dmnd
+
+[markers]
+# Marker sets should be in the format of kingdom.marker_set_name.[cutoffs,hmm]
+bacteria_single_copy = ${common:markers_dir}/bacteria.single_copy.hmm
+bacteria_single_copy_cutoffs = ${common:markers_dir}/bacteria.single_copy.cutoffs
+archaea_single_copy = ${common:markers_dir}/archaea.single_copy.hmm
+archaea_single_copy_cutoffs = ${common:markers_dir}/archaea.single_copy.cutoffs
