+++ databases/markers/bacteria.single_copy.cutoffs
@@ -1,0 +1,139 @@
+PF03485 32.55
+PF03484 26.55
+PF01121 84.65
+PF03772 82.7
+PF03602 38.8
+PF06418 208
+PF02224 79.35
+PF00712 48.7
+PF02767 49.65
+PF02768 44.5
+PF00035 20.2
+PF00889 108.35
+PF01176 39.25
+PF00113 206.35
+PF03952 92.45
+PF06574 73.25
+PF03147 39.75
+PF01687 59.3
+PF02938 40.05
+PF02527 77.85
+PF00958 61.3
+PF01025 72.25
+PF01018 98.9
+PF11987 61.7
+PF04760 29.95
+PF00707 51.55
+PF05198 47.95
+PF01715 129.15
+PF06421 79.7
+PF01795 174.05
+PF02873 46.15
+PF08529 61.55
+PF02410 44.9
+PF01195 99.15
+PF01252 61.35
+PF00162 236
+PF02912 33.55
+PF03726 25.6
+PF01416 42.45
+PF02033 44.9
+PF00154 276.55
+PF02132 21.25
+PF00825 40.55
+PF00687 75.1
+PF00466 37.3
+PF00298 42.65
+PF03946 45.9
+PF00542 44
+PF00572 81.3
+PF00238 79.55
+PF00252 77.45
+PF01196 55.75
+PF00828 37.55
+PF00861 54.85
+PF01245 71.7
+PF00181 52.45
+PF03947 87.3
+PF00453 70.5
+PF00829 53.45
+PF00237 57.6
+PF00276 37.45
+PF01016 55.4
+PF00830 36.8
+PF00831 33.7
+PF00297 66.35
+PF01783 28.2
+PF01632 29.5
+PF00573 99.95
+PF00281 34.55
+PF00673 61.25
+PF00347 53.25
+PF03948 37.95
+PF01281 30.55
+PF00338 58.05
+PF00411 73.45
+PF00164 84.3
+PF00416 56.2
+PF00312 41.7
+PF00886 36.25
+PF00366 41.3
+PF01084 36.3
+PF00203 54.9
+PF00318 133.15
+PF01649 35.95
+PF00189 49.5
+PF00163 32.9
+PF00333 42.55
+PF03719 44.9
+PF01250 45.2
+PF00177 95.55
+PF00410 72.35
+PF00380 70.4
+PF01782 30.65
+PF01000 34.3
+PF03118 38.9
+PF01193 34.35
+PF04997 157.25
+PF00623 79.8
+PF04983 41.85
+PF05000 24.6
+PF04998 119.4
+PF04563 39.95
+PF04561 37.15
+PF04565 45.6
+PF10385 36.6
+PF00562 258.45
+PF04560 46.45
+PF01765 101.65
+PF07499 16.6
+PF01330 26.85
+PF05491 49.5
+PF02773 108.9
+PF02772 77.05
+PF00584 27.5
+PF03840 29.2
+PF00344 180.7
+PF02403 43.15
+PF01668 43.4
+PF02978 52.8
+PF00763 62.2
+PF02882 103.7
+PF00121 140.2
+PF08275 58.6
+PF03461 40.05
+PF05698 55.15
+PF05697 61.8
+PF01746 80.75
+PF00750 120.85
+PF01409 161.15
+PF01509 83.55
+PF00627 13.3
+PF02130 59.35
+PF02367 46.1
+PF03652 60.75
+PF12344 33.75
+PF08459 81.9
+PF10458 26.5
+PF06071 61.9
+PF06689 31.7
