--- single-copy_markers/Archaea_single_copy_cutoffs.txt
@@ -1,162 +1,0 @@
-PF00113 150.75
-PF00121 39.45
-PF00137 36.2
-PF00162 181.15
-PF00164 60.25
-PF00177 65.4
-PF00181 20.35
-PF00189 24.9
-PF00203 39.75
-PF00237 51
-PF00238 55
-PF00252 46.05
-PF00253 23.25
-PF00276 23.7
-PF00281 25.75
-PF00297 114
-PF00298 31.7
-PF00312 28.8
-PF00318 55.9
-PF00327 24.05
-PF00333 37.9
-PF00334 88.05
-PF00344 145.45
-PF00347 33.25
-PF00366 37.65
-PF00368 235
-PF00380 50.5
-PF00410 37
-PF00411 64.6
-PF00416 47.5
-PF00466 32.4
-PF00562 204.75
-PF00572 39.35
-PF00573 63.9
-PF00623 107.9
-PF00670 81.05
-PF00673 37.65
-PF00679 34.25
-PF00687 67.15
-PF00709 130.9
-PF00736 29.55
-PF00749 120.7
-PF00750 78.1
-PF00752 30.15
-PF00827 128.6
-PF00831 25.65
-PF00832 28.85
-PF00833 24.2
-PF00861 51.85
-PF00867 39.9
-PF00900 40.7
-PF00919 31.45
-PF00935 41.9
-PF00958 47.05
-PF01000 40.2
-PF01015 111.55
-PF01035 38.1
-PF01090 84.1
-PF01092 48.7
-PF01139 244
-PF01157 40.35
-PF01172 38.7
-PF01174 82.5
-PF01180 71.95
-PF01191 41.75
-PF01192 23.5
-PF01194 41.25
-PF01198 35.6
-PF01200 44.25
-PF01201 53.6
-PF01246 30.9
-PF01253 34.75
-PF01269 143.65
-PF01280 85.95
-PF01282 36
-PF01287 20.65
-PF01351 67.7
-PF01496 85.05
-PF01509 31.5
-PF01599 28.6
-PF01655 57.3
-PF01667 31
-PF01725 78.25
-PF01780 46.15
-PF01798 80.45
-PF01813 94.35
-PF01849 24.05
-PF01864 76.25
-PF01866 149.75
-PF01868 37.8
-PF01884 47.8
-PF01896 25.85
-PF01907 34.15
-PF01912 104.05
-PF01920 34.45
-PF01922 35.95
-PF01941 285.4
-PF01948 54.45
-PF01949 92.65
-PF01951 62.7
-PF01967 75
-PF01973 46.8
-PF01981 64.6
-PF01982 62.9
-PF01984 43.1
-PF01990 26.2
-PF01991 55.45
-PF02005 144.1
-PF02006 119.55
-PF02748 25.95
-PF02834 40.65
-PF02938 22.75
-PF02978 46.7
-PF02996 34.95
-PF03439 32.45
-PF03484 25.45
-PF03719 35.45
-PF03764 46.75
-PF03853 68.05
-PF03874 17.45
-PF03876 25.4
-PF03911 15.55
-PF03946 35.5
-PF03947 67.75
-PF03950 37.35
-PF04010 32.85
-PF04019 56.65
-PF04104 51.55
-PF04127 79.55
-PF04135 35.05
-PF04406 32.45
-PF04414 104.55
-PF04560 42.4
-PF04561 45.9
-PF04563 89.55
-PF04565 39.65
-PF04566 32.5
-PF04567 21.4
-PF04919 104.9
-PF04981 84.6
-PF04983 53.05
-PF04997 142.85
-PF05000 48.5
-PF05221 105.15
-PF05670 39.6
-PF05746 46.9
-PF06026 88.6
-PF06093 18.4
-PF06418 205
-PF06463 40.05
-PF07541 41.75
-PF08068 28.05
-PF08069 36.15
-PF08071 21.25
-PF08608 30.8
-PF08704 38.5
-PF09173 46.35
-PF09239 100.4
-PF09249 66.75
-PF09377 31.2
-PF10559 14.9
-PF11987 34.05
